Remove unused experiment modules
